Second Commit

+++ src/main/java/com/dfc/network/repository/UserBinaryInfoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserBinaryInfoRepository extends JpaRepository<UserBinaryInfo, Long> {
+public interface UserBinaryInfoRepository extends JpaRepository<UserBinaryInfo, Integer> {
     UserBinaryInfo findByParentIdAndPosition(Integer parentId, String position);
     List<UserBinaryInfo> findByParentId(Integer parentId);
     UserBinaryInfo findByUserUserId(Integer userId);
